+++ lib/src/path_painter_builder.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import 'debug.dart';
 import 'line_animation.dart';
 import 'painter.dart';